# Optimize qualified identifier creation