## Description

A micro-optimization:
- Simply return the identifier, if there are no containers.
- Avoid appending by creating the identifier-array at the lowest level of the chain (when the size is known).
- Calculate the buffer size on the fly, and grow the string-builder buffer at once.

```
name                                                  old time/op    new time/op    delta
QualifiedIdentifierCreation/One_level-12              24.6ns ± 4%     2.0ns ± 1%   -91.96%  (p=0.008 n=5+5)
QualifiedIdentifierCreation/Three_levels-12            197ns ± 4%     105ns ± 7%   -46.93%  (p=0.008 n=5+5)

name                                                  old alloc/op   new alloc/op   delta
QualifiedIdentifierCreation/One_level-12               8.00B ± 0%     0.00B       -100.00%  (p=0.008 n=5+5)
QualifiedIdentifierCreation/Three_levels-12             120B ± 0%       64B ± 0%   -46.67%  (p=0.008 n=5+5)

name                                                  old allocs/op  new allocs/op  delta
QualifiedIdentifierCreation/One_level-12                1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
QualifiedIdentifierCreation/Three_levels-12             4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.008 n=5+5)

```

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels