@turbolent 2021-05-26
This is a great optimization! This code is called often, so optimizing it makes a lot of sense 👍